server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
numpy==1.26.4
requests==2.31.0
# Pipecat dependencies - using more recent versions
openai==1.35.0
anthropic==0.28.0
deepgram-sdk==3.2.7
# Audio processing
pyaudio==0.2.11
scipy==1.11.4
